Rename package for com.stripe.android.stripecardscan.scanui.util
